+++ EFCore.NamingConventions.Test/LowerCaseNamingTest.cs
@@ -1,3 +1,4 @@
+﻿using System.Globalization;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
@@ -14,6 +15,9 @@ namespace EFCore.NamingConventions.Test
             Assert.Equal("simpleblog", entityType.GetTableName());
         }
 
+
+    
+
         [Fact]
         public void Column_name_is_rewritten()
         {
@@ -21,8 +25,23 @@ namespace EFCore.NamingConventions.Test
             var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
             Assert.Equal("id", entityType.FindProperty("Id").GetColumnName());
             Assert.Equal("fullname", entityType.FindProperty("FullName").GetColumnName());
+        }   
+        [Fact]
+        public void Column_name_is_rewritten_in_turkish()
+        {
+            using var context = CreateContext(CultureInfo.CreateSpecificCulture("tr_TR"));
+            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
+            Assert.Equal("ıd", entityType.FindProperty("Id").GetColumnName());
+            Assert.Equal("fullname", entityType.FindProperty("FullName").GetColumnName());
         }
-
+        [Fact]
+        public void Column_name_is_rewritten_in_invariant()
+        {
+            using var context = CreateContext(CultureInfo.InvariantCulture);
+            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
+            Assert.Equal("id", entityType.FindProperty("Id").GetColumnName());
+            Assert.Equal("fullname", entityType.FindProperty("FullName").GetColumnName());
+        }
         [Fact]
         public void Primary_key_name_is_rewritten()
         {
@@ -55,6 +74,6 @@ namespace EFCore.NamingConventions.Test
             Assert.Equal("ix_simpleblog_fullname", entityType.GetIndexes().Single().GetName());
         }
 
-        TestContext CreateContext() => new TestContext(NamingConventionsExtensions.UseLowerCaseNamingConvention);
+        TestContext CreateContext(CultureInfo culture = null) => new TestContext((builder) => builder.UseLowerCaseNamingConvention(culture));
     }
 }
