--- EFCore.NamingConventions.Test/LowerCaseNamingWithCultureTest.cs
@@ -1,61 +1,0 @@
-﻿using System.Globalization;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using Xunit;
-
-namespace EFCore.NamingConventions.Test
-{
-    public class LowerCaseNamingWithCultureTest : RewriterTestBase
-    {
-        [Fact]
-        public void Table_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("simpleblog", entityType.GetTableName());
-        }
-
-        [Fact]
-        public void Column_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("ıd", entityType.FindProperty("Id").GetColumnName());
-            Assert.Equal("fullname", entityType.FindProperty("FullName").GetColumnName());
-        }
-
-        [Fact]
-        public void Primary_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("pk_simpleblog", entityType.GetKeys().Single(k => k.IsPrimaryKey()).GetName());
-        }
-
-        [Fact]
-        public void Alternative_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("ak_simpleblog_somealternativekey", entityType.GetKeys().Single(k => !k.IsPrimaryKey()).GetName());
-        }
-
-        [Fact]
-        public void Foreign_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(Post));
-            Assert.Equal("fk_post_simpleblog_blogıd", entityType.GetForeignKeys().Single().GetConstraintName());
-        }
-
-        [Fact]
-        public void Index_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("ıx_simpleblog_fullname", entityType.GetIndexes().Single().GetName());
-        }
-
-        TestContext CreateContext() => new TestContext(NamingConventionsExtensions.UseLowerCaseNamingConvention, CultureInfo.CreateSpecificCulture("tr_TR"));
-    }
-}
