+++ EFCore.NamingConventions.Test/UpperCaseNamingTest.cs
@@ -1,3 +1,4 @@
+﻿using System.Globalization;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using Xunit;
@@ -14,6 +15,21 @@ namespace EFCore.NamingConventions.Test
             Assert.Equal("SIMPLEBLOG", entityType.GetTableName());
         }
 
+        [Fact]
+        public void Table_name_is_rewritten_in_turkish()
+        {
+            using var context = CreateContext(CultureInfo.CreateSpecificCulture("tr_TR"));
+            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
+            Assert.Equal("SİMPLEBLOG", entityType.GetTableName());
+        }
+
+        [Fact]
+        public void Table_name_is_rewritten_in_invariant()
+        {
+            using var context = CreateContext(CultureInfo.InvariantCulture);
+            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
+            Assert.Equal("SIMPLEBLOG", entityType.GetTableName());
+        }
         [Fact]
         public void Column_name_is_rewritten()
         {
@@ -54,7 +70,6 @@ namespace EFCore.NamingConventions.Test
             var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
             Assert.Equal("IX_SIMPLEBLOG_FULLNAME", entityType.GetIndexes().Single().GetName());
         }
-
-        TestContext CreateContext() => new TestContext(NamingConventionsExtensions.UseUpperCaseNamingConvention);
+        TestContext CreateContext(CultureInfo culture = null) => new TestContext((builder) => builder.UseUpperCaseNamingConvention(culture));
     }
 }
