--- EFCore.NamingConventions.Test/UpperCaseNamingWithCultureTest.cs
@@ -1,61 +1,0 @@
-﻿using System.Globalization;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
-using Xunit;
-
-namespace EFCore.NamingConventions.Test
-{
-    public class UpperCaseNamingWithCultureTest : RewriterTestBase
-    {
-        [Fact]
-        public void Table_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("SİMPLEBLOG", entityType.GetTableName());
-        }
-
-        [Fact]
-        public void Column_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("ID", entityType.FindProperty("Id").GetColumnName());
-            Assert.Equal("FULLNAME", entityType.FindProperty("FullName").GetColumnName());
-        }
-
-        [Fact]
-        public void Primary_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("PK_SİMPLEBLOG", entityType.GetKeys().Single(k => k.IsPrimaryKey()).GetName());
-        }
-
-        [Fact]
-        public void Alternative_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("AK_SİMPLEBLOG_SOMEALTERNATİVEKEY", entityType.GetKeys().Single(k => !k.IsPrimaryKey()).GetName());
-        }
-
-        [Fact]
-        public void Foreign_key_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(Post));
-            Assert.Equal("FK_POST_SİMPLEBLOG_BLOGID", entityType.GetForeignKeys().Single().GetConstraintName());
-        }
-
-        [Fact]
-        public void Index_name_is_rewritten()
-        {
-            using var context = CreateContext();
-            var entityType = context.Model.FindEntityType(typeof(SimpleBlog));
-            Assert.Equal("IX_SİMPLEBLOG_FULLNAME", entityType.GetIndexes().Single().GetName());
-        }
-
-        TestContext CreateContext() => new TestContext(NamingConventionsExtensions.UseUpperCaseNamingConvention, CultureInfo.CreateSpecificCulture("tr_TR"));
-    }
-}
